Switch to u8 for data

+++ src/utils.rs
@@ -3,10 +3,10 @@ pub fn mod_floor(a: i64, m: i64) -> i64 {
     ((a % m) + m) % m
 }
 
-pub fn ceil_log2(x: i64) -> i64 {
-    (x as f64).log2().ceil() as i64
+pub fn ceil_log2(x: u8) -> u8 {
+    (x as f64).log2().ceil() as u8
 }
 
-pub fn power_of_two(num: i64) -> bool {
+pub fn power_of_two(num: u8) -> bool {
     num > 0 && (num & (num - 1)) == 0
 }
